Add excludeListType to jck9/runtime.jti
This appears not to cause a problem on @lumpfish's system
but does on mine - reason unknown - but the directive is in
the jck8 runtime.jti and also the jck9 compiler.jti and devtools.jti
so it doesn't seem unreasonable to have it in there.

JCK stderr + set jck.excludeList.customFiles "/home/cloudusr/git/jck9/excludes/jck9.jtx "
JCK stderr Configuration file can not be imported due to question with specified tag "jck.excludeList.customFiles" was not found on path. The path is:jck.intro